requirements.txt
streamlit
googel-generativeai
python-dotenv
